fix: rename profile-modal.js to fit Windows wheel path budget
