+++ src/reachy_mini_conversation_app/static_v2/js/views/home.js
@@ -28,7 +28,7 @@ import {
   avatarFor,
 } from "../constants.js";
 import { $, clear, h, prettifyProfileName } from "../ui.js";
-import { openCustomProfileModal } from "../components/custom-profile-modal.js";
+import { openCustomProfileModal } from "../components/profile-modal.js";
 
 /**
  * @param {{ outlet: HTMLElement, signal: AbortSignal, navigate: (route: string) => void }} ctx
